% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factorial}
\alias{ass}
\alias{factorial}
\title{ASS(ign) survey participants :)}
\usage{
factorial(n)

ass(n)
}
\arguments{
\item{n}{- integer, number of participants in control or treatment group (both have equal size)}
}
\value{
integer matrix, with all possible 0-1 participants assignments.
}
\description{
This function generates all possible 0-1 assignment vectors of 2*n survey participants
in such a way that exactly n of them are assigned to group 0 (control)
and the other n ones are assigned to group 1 (treatment). To simplyfy this task I have used
std::next_permutation() - this suggestion I have found on stackoverflow.
}

