% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hw6package-package.R
\docType{package}
\name{hw6package-package}
\alias{hw6package-package}
\title{hw6package}
\description{
This package consists of solutions to task 1, 2 and 3 included in homework 6.
Solution functions are called mode, simplify2array and ass.
This is my first R package :)
}

