% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{Simple function calculating mode.}
\usage{
mode(x)
}
\arguments{
\item{integer}{vector}
}
\value{
mode - single integer number -  the mode
}
\description{
This function, given the input integer vector, calculates mode - the most
frequently occurring value. If the mode is ambiguous it returns last one.
}

