% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array}
\alias{simplify2array}
\title{List of vectors ==> Matrix}
\usage{
simplify2array(vec_list)
}
\arguments{
\item{vec_list}{- list of input numeric vectors}
}
\value{
numeric matrix, where number of rows equals the length of the list and number of columns
equal the length of vectors in the list.
}
\description{
This function changes input list of numeric vectors to a matrix.
We assume that vectors have equal length.
}

